task2/requirements.txt
aiocsv==1.2.5
aiodns==3.1.1
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
environs==10.3.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
marshmallow==3.20.2
multidict==6.0.4
packaging==23.2
psycopg2-binary==2.9.9
pycares==4.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
soupsieve==2.5
SQLAlchemy==2.0.25
typing_extensions==4.9.0
yarl==1.9.4
